split out odds/ends and misc settings into separate files
