+++ source/views/settings/sections/odds-and-ends.js
@@ -6,9 +6,6 @@ import type {TopLevelViewPropsType} from '../../types'
 import {setFeedbackStatus} from '../../../flux/parts/settings'
 import {connect} from 'react-redux'
 import {CellToggle} from '../../components/cells/toggle'
-import {PushButtonCell} from '../../components/cells/push-button'
-import {trackedOpenUrl} from '../../components/open-url'
-import * as Icons from '@hawkrives/react-native-alternate-icons'
 import {sectionBgColor} from '../../components/colors'
 
 type Props = TopLevelViewPropsType & {
@@ -16,58 +13,10 @@ type Props = TopLevelViewPropsType & {
 	feedbackDisabled: boolean,
 }
 
-type State = {
-	supported: boolean,
-}
-
-class OddsAndEndsSection extends React.PureComponent<Props, State> {
-	state = {
-		supported: false,
-	}
-
-	componentDidMount() {
-		this.checkIfCustomIconsSupported()
-	}
-
-	checkIfCustomIconsSupported = async () => {
-		const supported = await Icons.isSupported()
-		this.setState(() => ({supported}))
-	}
-
-	onPressButton = (id: string) => {
-		this.props.navigation.navigate(id)
-	}
-
-	onCreditsButton = () => this.onPressButton('CreditsView')
-	onPrivacyButton = () => this.onPressButton('PrivacyView')
-	onLegalButton = () => this.onPressButton('LegalView')
-	onSourceButton = () =>
-		trackedOpenUrl({
-			url: 'https://github.com/StoDevX/AAO-React-Native',
-			id: 'ContributingView',
-		})
-	onAppIconButton = () => this.onPressButton('IconSettingsView')
-
+class OddsAndEndsSection extends React.PureComponent<Props> {
 	render() {
 		return (
 			<React.Fragment>
-				<Section header="MISCELLANY" sectionTintColor={sectionBgColor}>
-					{this.state.supported ? (
-						<PushButtonCell
-							onPress={this.onAppIconButton}
-							title="Change App Icon"
-						/>
-					) : null}
-
-					<PushButtonCell onPress={this.onCreditsButton} title="Credits" />
-					<PushButtonCell
-						onPress={this.onPrivacyButton}
-						title="Privacy Policy"
-					/>
-					<PushButtonCell onPress={this.onLegalButton} title="Legal" />
-					<PushButtonCell onPress={this.onSourceButton} title="Contributing" />
-				</Section>
-
 				<Section header="ODDS &amp; ENDS" sectionTintColor={sectionBgColor}>
 					<Cell cellStyle="RightDetail" detail={version} title="Version" />
